test(entreprise inclusive): add e2e test and remove stub
