+++ clients/_test/clientSearchRechercheEntreprise/_snapshots/search-533744991.json
@@ -1,0 +1,147 @@
+{
+  "args": {
+    "pageResultatsRecherche": 1,
+    "searchTerms": "533744991",
+    "searchFilterParams": ""
+  },
+  "result": {
+    "results": [
+      {
+        "siren": "533744991",
+        "nom_complet": "IDDEE 13",
+        "nom_raison_sociale": "IDDEE 13",
+        "sigle": null,
+        "nombre_etablissements": 3,
+        "nombre_etablissements_ouverts": 1,
+        "siege": {
+          "activite_principale": "38.32Z",
+          "activite_principale_registre_metier": null,
+          "annee_tranche_effectif_salarie": "2022",
+          "adresse": "ZONE ARTISANORD 1 AVENUE DE LA BAUXITE 13015 MARSEILLE",
+          "caractere_employeur": "O",
+          "cedex": null,
+          "code_pays_etranger": null,
+          "code_postal": "13015",
+          "commune": "13215",
+          "complement_adresse": "ZONE ARTISANORD",
+          "coordonnees": "43.338657,5.367455",
+          "date_creation": "2017-11-15",
+          "date_debut_activite": "2017-11-15",
+          "date_fermeture": null,
+          "date_mise_a_jour": null,
+          "date_mise_a_jour_insee": "2024-03-30T07:54:53",
+          "departement": "13",
+          "distribution_speciale": null,
+          "epci": "200054807",
+          "est_siege": true,
+          "etat_administratif": "A",
+          "geo_adresse": "1 Avenue de la bauxite 13015 Marseille",
+          "geo_id": "13215_0890_00001",
+          "indice_repetition": null,
+          "latitude": "43.338657",
+          "libelle_cedex": null,
+          "libelle_commune": "MARSEILLE",
+          "libelle_commune_etranger": null,
+          "libelle_pays_etranger": null,
+          "libelle_voie": "DE LA BAUXITE",
+          "liste_enseignes": null,
+          "liste_finess": null,
+          "liste_id_bio": null,
+          "liste_idcc": [
+            "0637"
+          ],
+          "liste_id_organisme_formation": null,
+          "liste_rge": null,
+          "liste_uai": null,
+          "longitude": "5.367455",
+          "nom_commercial": null,
+          "numero_voie": "1",
+          "region": "93",
+          "siret": "53374499100035",
+          "statut_diffusion_etablissement": "O",
+          "tranche_effectif_salarie": "21",
+          "type_voie": "AVENUE"
+        },
+        "activite_principale": "38.32Z",
+        "categorie_entreprise": "PME",
+        "caractere_employeur": null,
+        "annee_categorie_entreprise": "2022",
+        "date_creation": "2011-07-15",
+        "date_fermeture": null,
+        "date_mise_a_jour": "2024-12-05T19:05:38",
+        "date_mise_a_jour_insee": "2024-03-22T14:26:06",
+        "date_mise_a_jour_rne": "2024-05-19T16:49:33",
+        "dirigeants": [
+          {
+            "nom": "PELLUAULT",
+            "prenoms": "ALEXIS MARIE BRUNO",
+            "annee_de_naissance": "1969",
+            "date_de_naissance": "1969-06",
+            "qualite": "Président de SAS",
+            "nationalite": "Française",
+            "type_dirigeant": "personne physique"
+          }
+        ],
+        "etat_administratif": "A",
+        "nature_juridique": "5710",
+        "section_activite_principale": "E",
+        "tranche_effectif_salarie": "21",
+        "annee_tranche_effectif_salarie": "2022",
+        "statut_diffusion": "O",
+        "matching_etablissements": [],
+        "immatriculation": {
+          "date_debut_activite": "2011-07-15",
+          "date_immatriculation": "2011-08-04",
+          "date_radiation": null,
+          "duree_personne_morale": 99,
+          "nature_entreprise": [
+            "Gestion de biens",
+            "Commerciale"
+          ],
+          "date_cloture_exercice": "3112",
+          "capital_social": 10000,
+          "capital_variable": false,
+          "devise_capital": "EUR",
+          "indicateur_associe_unique": false
+        },
+        "finances": {
+          "2021": {
+            "ca": 0,
+            "resultat_net": 86584
+          }
+        },
+        "complements": {
+          "collectivite_territoriale": null,
+          "convention_collective_renseignee": true,
+          "liste_idcc": [
+            "0637"
+          ],
+          "egapro_renseignee": false,
+          "est_association": false,
+          "est_bio": false,
+          "est_entrepreneur_individuel": false,
+          "est_entrepreneur_spectacle": false,
+          "est_ess": false,
+          "est_finess": false,
+          "est_organisme_formation": false,
+          "est_qualiopi": false,
+          "liste_id_organisme_formation": null,
+          "est_rge": false,
+          "est_service_public": false,
+          "est_siae": true,
+          "est_societe_mission": false,
+          "est_uai": false,
+          "identifiant_association": null,
+          "statut_entrepreneur_spectacle": null,
+          "type_siae": "EI"
+        },
+        "slug": "iddee-13-533744991",
+        "dateDerniereMiseAJour": "2023-09-21T03:34:50"
+      }
+    ],
+    "total_results": 1,
+    "page": 1,
+    "per_page": 10,
+    "total_pages": 1
+  }
+}
