+++ cypress/mocks/routes.ts
@@ -1,6 +1,7 @@
 import routes from '#clients/routes';
 import { http } from 'msw';
 import { apiBioHandler } from './handlers/api-bio';
+import { apiInclusionHandler } from './handlers/api-inclusion';
 import { eoriHandler } from './handlers/eori';
 import { rechercheEntrepriseHandler } from './handlers/recherche-entreprises';
 import { rgeHandler } from './handlers/rge';
@@ -15,4 +16,8 @@ export const routesHandlers = [
   ),
   http.get(`${routes.certifications.rge.api}`, rgeHandler),
   http.get(`${routes.certifications.bio.api}`, apiBioHandler),
+  http.get(
+    routes.certifications.entrepriseInclusive.api.siren,
+    apiInclusionHandler
+  ),
 ];
